[Driver] modify saveobj and loadobj to recreate a driver
Instead of just saving the pointer to the C++ object when the
matlab wrapper is saved to disk, we store the constructor argument
along with any object that was used to modify the state of the driver.
When a driver is loaded, we create a new driver from the saved
contrutor arguments and reset the state if necessary.