Change tensor representation
The tensors used to be represented as a 3x(3*#elements) matrix, i.e.
a concatenation of the 3x3 tensors. In order to simplify the handling
of the tensors and use the representation that is already used by
simbio, we change the representation to 9x(#elements) where each tensor
is stored row-wise.